+++ src/main/java/webserver/load/LoadResult.java
@@ -1,0 +1,4 @@
+package webserver.load;
+
+public record LoadResult(byte[] content, String path) {
+}
